some fix, ad getDto

+++ src/main/java/telran/java48/person/dto/CityPopulationDto.java
@@ -4,10 +4,11 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
 public class CityPopulationDto {
 	String city;
-	Integer population;
+	Long population;
 }
